Patch from 'fsck' to fix rtsp* modes
It fixes rtsp* modes to for audio (and additionally video). It also
attempts to synthesize the rtsp urls from all rtmp connection mode data.
mplayer used to work for rtsp audio modes but no longer... Seems that
now only vlc works to download/stream audio. Only audio is decodable by
vlc from video streams. This is a vlc limitation but the video stream is
in there :-) - just not recognised.

Maybe someone in the live555 team will fix it to recognise the video
stream ID: "video/X-HX-AVC1" ??