Update iPlayer EMP version
Some time back, the BBC updated to a new version of its Flash embedded
media player for the iPlayer.  I don't know of any functional changes
that affect get_iplayer, but we need to update the script to pass the
new URL and ensure that SWF verification calculates the hashes of the
correct file.

Fixes download of HD content from iPlayer.  Users may need to remove the
.swfinfo file where rtmpdump caches SWF verification information.