Show title correctly in terminal or tmux
When running get_iplayer in a terminal or tmux session the title of the 
window will show as "perl". The attached one-line change - too small a 
change to call a "patch", really ;) - will alter behaviour so that 
"get_iplayer" is displayed instead, enabling the window in which 
get_iplayer is running to be more readily identified.

See also: http://www.perlmonks.org/index.pl?node_id=344941

aB.