Output AAC as M4A for iTunes with metadata tags
If the downloaded file is from a flashaac mode then a second
pass of ffmpeg removes the ADTS header and remuxes to M4A
container. Two passes are required as direct FLV to M4A
results in a broken M4A which will not play and reports
incorrect duration. FLV -> AAC -> M4A works. AtomicParsley
is then used to correctly tag the M4A file