Ensure safe 32-bit date if timegm() fails
get_iplayer uses timegm() to calculate expiry dates for
downloaded programmes.  In 32-bit versions of perl or any
perl < 5.10.1, timegm() croaks for any date greater than
2038-01-16 23:59:59.  The current approach of setting
year = 2038 for any dates in 2038 and after is insufficient
to ensure a safe 32-bit date.  This patch changes the code
to run timegm() within an eval block.  If timegm() croaks,
epoch date is set to arbitrary safe value (2038-01-01 00:00:00).
Redundant calls to to timegm() have also been removed.

The problem corrected by this patch is very much an edge case,
so far only observed when using a hacked get_iplayer to download
media clips embedded in other BBC web sites.  However, the logic
is still incorrect, so this patch has been created to fix it.