Ensure output file extension changes with download mode
If a download fails to start, e.g., because iPlayer metadata is bad,
get_iplayer will try the next available mode that matches the user's
mode preferences.  However, even if the second mode produces an output
file with a different format, the output file will be written with the
file extension appropriate for the first, failed mode.  For example, with:
radiomode=flashaudio,flashaac if the flashaudio stream fails to start,
the flashaac stream will be used.  However, the resulting AAC file will
still be written with a .mp3 extension.  This patch checks for changes
to the extension of output files and ensures the file name is altered
if necessary.