Use File::Spec functions for programme/thumbnail file paths
This change ensures that file paths used in substitution
parameters (<filename>, <filepart>, <thumbfile>) are:
a) created in an OS-independent manner; b) not sent through
normal escaping when expanded (since they should now contain
only safe characters).  This is primarily for the benefit of
Windows/DOS system commands (e.g., COPY) that may be confused
by mixed or escaped directory separators in a file path when
employed as part of a user command.  Calls to File::Spec
functions are fully-qualified because File::Spec::Functions
is not available in Windows installer version of get_iplayer.