Revamped --showoptions and search options initialisation
The init_search() subroutine has been changed:
1. An unnecessary search will no longer be performed when
--showoptions is used
2. All options processing occurs at head of init_search() so that
--showoptions may be invoked in a short-circuit fashion.
3. Validation error will be shown along with options listing
when --showoptions used
4. --{metadata,subs,thumb}only will force --nowrite, though it may
still be specified on the command line.  Programme may still be
streamed with --player, --nowrite, and --stdout when --get is used.

The effect of these changes is that --showoptions can be added
to any command line as means to check options before search
execution or in event of an error.