Update --versions help text and error message for signed-only programmes
When attempting to download signed- or audiodescribed-only programmes
without explicitly adding 'signed' or 'audiodescribed' to --versions,
get_iplayer would produce the slightly cryptic error message:

WARNING: No programmes are available for this pid

The error message has been expanded to indicate which versions are
available.  The help text for --versions has been expanded to explain
the list processions and to indicate that 'default' is the default value.