Web PVR: use dynamic discovery of configured ffmpeg
The location of ffmpeg may be changed during update by the Windows
installer and thus should not be hard-coded in get_iplayer.cgi.cmd.
This change enables web pvr to query get_iplayer for configured
ffmpeg if it is not specified on get_iplayer.cgi command line.  This
is probably only relevant on Windows, where ffmpeg will not generally
be in PATH.