Added support for optional prefixes in substitution parameters
Substitution parameters may now contain prefixes before the parameter
name. The prefix will not be printed if the associated token is empty
or undefined.  For example:

--fileprefix="<nameshort><-senum>"

will expand to "The_Last_Station" for the film of the same name since
<senum> is undefined for films.  For Top Gear series 19 episode 2,
where <senum> is defined, it will expand to "Top_Gear-s19e02".