Added support for iPlayer download links (bbc-ipd URL scheme)
get_iplayer can now accept URLs that use the bbc-ipd URI scheme.  This
scheme is used for download links on the iPlayer TV site (links that
begin with "bbc-ipd:download/"). If a bbc-ipd link contains "/hd/" and
--modes has not been specified, --modes=best is forced in order to get
an HD stream. Only one bbc-ipd link can be specified since this feature
is intended for use from a URL handler. If a system has a suitable
handler configured for the bbc-ipd scheme, it is possible to invoke
get_iplayer directly by clicking the "Download" link for TV programmes
on the iPlayer site. Configuration of URL handlers is specific to each
each OS, so get_iplayer does not implement any particular method. That
is left up to packagers and users.

Thanks to Peter Oliver for original PR

Closes #30