Make 404 errors non-fatal for HLS/DASH streaming of on-demand programmes
If a segment cannot be retrieved and a 404 error code is received, the
download will not be retried. An appropriate warning will be displayed
and the download sequence will continue if not at the end of the segment
list. It appears that the DASH manifest data is not sufficiently accurate
to calculate the correct number of segments for all programmes. This has
manifested in the wild as the segment list being 1 item too long. This
is only a DASH problem. HLS manifests already contain segment list, so
this problem is unlikely to arise with them.

Closes #237