Removed use of --versions in searches
This hasn't worked in a very long time, and it is possible to get
erroneous search results if --versions doesn't include "default".
Help text updated to indicate --versions is for recording only. The
equivalent Programme Version field in Web PVR has been moved to the
Recordings tab. More work will be necessary to completely remove all
vestiges, but ignoring --versions in search should be enough for now.