HLS is now default format for all programme types
flash modes for radio/tv and shoutcast modes for live radio are
deprecated, but still available. However, default processing does
not fall back to flash or shoutcast if hls/hvf is not available.
Currently, the built-in HLS/DASH streamer is used for all programme
types. However, streamed and live programmes may be handled solely
by ffmpeg in the future.

Closes #130, #131