Fix schedule refresh with limit at end of year
When a refresh limit is set and not divisible by 7 days, refreshing programme schedules at the end of the year can incorrectly try to download schedules for all weeks in at least the next year. I detected this problem today because my refresh limit is set to 30 days and get_iplayer attempted to download schedules for all weeks of 2017. This seems to be because today is in "week 52" but starting from 30 days ago and incrementing 7 days at a time means that the Sunday that is in "this week" falls into what Perl calls "week 0" of 2017.

The earliest date I've found in my non-scientific search where this behaviour becomes a significant problem is December 31st 2018, which is a Monday and "week 53". In this case, for users with `$limit_days + 1` divisible by 7 get_iplayer will attempt to download schedules for all weeks to the end of 2024. More worrying, for users with neither `$limit_days` nor `$limit_days + 1` divisible by 7 (73% of the possible non-zero refresh limits), get_iplayer will go into what looks like an infinite loop generating the list of schedule weeks it wants to download. I'm sure it could be proven this will be infinite with maths, testing with limits of 1-5 days reached a year of 20000+ without matching the loop escape criteria.