Trim subtitles according to --start and --stop
Actual start and stop times (to nearest chunk boundary) are used to trim
subtitles when --start and --stop are specified. Subtitles before start
time and after stop time are skipped when producing the .srt subtitles
file. The value of --suboffset is taken into account. Does not apply to
--subsraw. Also does not apply to --subtitles-only because actual start
and stop times are not known unless media file is downloaded.