Implemented --pvr-series option
Use --pvr-series to create a name-based PVR search for each unique series
name in search results. It is best to be conservative and use a single
index number or episode PID in "pid:<PID>" form as search term. There is
no need to specify a PVR search name. One will be generated. This option
is analogous to the "Add Series" links in Web PVR search results. Cannot
be used with --pid.