Ensure cache is always topped up to 30 days
Whenever cache is refreshed, it is now topped up to 30 days. By default,
refresh will go back to value of the timestamp of the most recently added
record minus the value of --expiry (default 14400) as overlap. The mod
time of the cache file is no longer used as the default basis for calculation.
This means that the value returned by --show-cache-age, which relies on
cache file mod time, may underestimate the cache age.

Use --refresh-limit to restrict how far back refresh goes in programme
schedules to update the cache.

Removed --cache-reset option and removed some extra white space.