revamped recording modes, 50fps now default
- Removed all "hls"-prefix tv and radio modes, except for "hlshd"
(1280x720@25fps). Although "hlshd" appears to be on the way out, some
older programmes may still be available in that format.

- 50fps streams will now be tried first for 1280x720, 960x540, and 704x396
video, with 25fps streams as backup for 960x540 and 704x396. This means
that those downloads will double in size compared to previous releases
when using default settings.

- removed --fps50 option as it is now the default

- added --fps25 option to force use of 25fps streams only (HD not available)

- added "Use only 25fps streams" option in Recording tab of Web PVR

- removed --no-hq-audio option (unused)

- removed "vgood" shortcuts. Now mapped to "better".

- removed "worse" shortcuts. Now mapped to "good".

- removed "vhigh" quality level. Now mapped to "xsd";

- removed "vhigh" quality level. Now mapped to "xsd";

- removed "std" quality level. Now mapped to "xhigh";

- "dash" and "hls" can now be used as shortcuts/prefixes and will be
mapped to corresponding modes for programme type

- Added hd,sd,high,low shortcuts for tv and high,std,med,low shortcuts
for radio. They may be prefixed with "tv" or "radio", respectively, for
use with --modes or Web PVR. These shortcuts provide a means to download
only streams of the corresponding resolution or bitrate, and may be
combined. Don't use only "hd" with audiodescribed programmes since HD
streams generally are not available. Don't use only "med" for tv programmes
or only "hd" or "sd" for radio programmes since such streams don't exist.
Don't use only "hd" with --fps25 since HD video is not available.
Be careful when using only "high" or "std" for World Service programmes
since such streams may only be available for podcast versions.

- HLS (haf) is now default format for radio programmes

- sort mode lists by quality, then format